testutil: move padded checker to separate file
